rm-dir: add some docs

+++ rm-dir/main.go
@@ -18,7 +18,13 @@ func main() {
 	flag.StringVar(&directories, "dirs", "", "comma seperated list of directories to remove")
 	flag.Parse()
 
-	for _, dir := range strings.Split(directories, ",") {
+	dirs := strings.Split(directories, ",")
+	if len(dirs) == 1 && dirs[0] == "" {
+		fmt.Println("at least 1 directory is required")
+		os.Exit(1)
+	}
+
+	for _, dir := range dirs {
 		args := []string{"filter-branch", "-f", "--index-filter", fmt.Sprintf("\"git rm --cached --ignore-unmatch -r %s\"", dir), "--prune-empty"}
 		cmd := exec.Command("git", args...)
